chore: move examples styles to shared data
